Move tree-agent types and expose funtions to new package
