+++ feeds/internal-build.txt
@@ -25,6 +25,7 @@
 @fluidframework/driver-utils
 @fluidframework/container-loader
 @fluidframework/undo-redo
+@fluidframework/tree-agent-types
 @fluidframework/tree-agent-ses
 @fluidframework/tree-agent-langchain
 @fluidframework/tree-agent
