+++ feeds/internal-test.txt
@@ -31,6 +31,7 @@
 @fluidframework/driver-utils
 @fluidframework/container-loader
 @fluidframework/undo-redo
+@fluidframework/tree-agent-types
 @fluidframework/tree-agent-ses
 @fluidframework/tree-agent-langchain
 @fluidframework/tree-agent
