+++ packages/dds/tree/package.json
@@ -176,6 +176,7 @@
 		"@fluidframework/runtime-utils": "workspace:~",
 		"@fluidframework/shared-object-base": "workspace:~",
 		"@fluidframework/telemetry-utils": "workspace:~",
+		"@fluidframework/tree-agent-types": "workspace:~",
 		"@sinclair/typebox": "^0.34.13",
 		"@tylerbu/sorted-btree-es6": "^2.1.1",
 		"@ungap/structured-clone": "^1.2.0",
