+++ packages/framework/tree-agent-langchain/src/test/domains/users.ts
@@ -7,10 +7,10 @@ import { SchemaFactoryAlpha } from "@fluidframework/tree/alpha";
 import {
 	buildFunc,
 	exposeMethodsSymbol,
+	typeFactory as tf,
 	type ExposedMethods,
 	type IExposedMethods,
-	typeFactory as tf,
-} from "@fluidframework/tree-agent/alpha";
+} from "@fluidframework/tree-agent-types/alpha";
 
 // eslint-disable-next-line @eslint-community/eslint-comments/disable-enable-pair
 /* eslint-disable jsdoc/require-jsdoc */
