+++ packages/framework/tree-agent-types/.npmignore
@@ -1,0 +1,7 @@
+nyc
+*.log
+**/*.tsbuildinfo
+src/test
+dist/test
+lib/test
+**/_api-extractor-temp/**
