+++ packages/framework/tree-agent-types/api-report/tree-agent-types.public.api.md
@@ -1,0 +1,7 @@
+## Public API Report File for "@fluidframework/tree-agent-types"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+```
