+++ packages/framework/tree-agent-types/package.json
@@ -1,12 +1,12 @@
 {
-	"name": "@fluidframework/tree-agent",
+	"name": "@fluidframework/tree-agent-types",
 	"version": "2.92.0",
-	"description": "Experimental package to simplify integrating AI into Fluid-based applications",
+	"description": "Shared types for exposing schema methods and properties to an LLM agent",
 	"homepage": "https://fluidframework.com",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/microsoft/FluidFramework.git",
-		"directory": "packages/framework/tree-agent"
+		"directory": "packages/framework/tree-agent-types"
 	},
 	"license": "MIT",
 	"author": "Microsoft and contributors",
@@ -110,13 +110,7 @@
 		"temp-directory": "nyc/.nyc_output"
 	},
 	"dependencies": {
-		"@anthropic-ai/sdk": "^0.39.0",
-		"@fluidframework/core-utils": "workspace:~",
-		"@fluidframework/runtime-utils": "workspace:~",
-		"@fluidframework/telemetry-utils": "workspace:~",
-		"@fluidframework/tree": "workspace:~",
-		"@fluidframework/tree-agent-types": "workspace:~",
-		"uuid": "^11.1.0"
+		"@fluidframework/telemetry-utils": "workspace:~"
 	},
 	"devDependencies": {
 		"@arethetypeswrong/cli": "^0.18.2",
@@ -126,17 +120,9 @@
 		"@fluidframework/build-common": "^2.0.3",
 		"@fluidframework/build-tools": "catalog:buildTools",
 		"@fluidframework/eslint-config-fluid": "catalog:eslint",
-		"@fluidframework/id-compressor": "workspace:~",
-		"@fluidframework/runtime-utils": "workspace:~",
-		"@fluidframework/test-runtime-utils": "workspace:~",
-		"@langchain/anthropic": "^0.3.24",
-		"@langchain/core": "^0.3.80",
-		"@langchain/google-genai": "^0.2.16",
-		"@langchain/openai": "^0.6.12",
 		"@microsoft/api-extractor": "7.52.11",
 		"@types/mocha": "^10.0.10",
 		"@types/node": "~20.19.30",
-		"@types/uuid": "^9.0.2",
 		"c8": "^10.1.3",
 		"concurrently": "^9.2.1",
 		"copyfiles": "^2.4.1",
@@ -148,7 +134,6 @@
 		"mocha-multi-reporters": "^1.5.1",
 		"prettier": "~3.6.2",
 		"rimraf": "^6.1.3",
-		"ses": "^1.14.0",
 		"typescript": "~5.4.5"
 	},
 	"typeValidation": {
