+++ packages/framework/tree-agent-types/src/test/tsconfig.json
@@ -1,0 +1,16 @@
+{
+	"extends": "../../../../../common/build/build-common/tsconfig.test.node16.json",
+	"compilerOptions": {
+		"rootDir": "./",
+		"outDir": "../../lib/test",
+		"types": ["node", "mocha"],
+		"declaration": true,
+		"declarationMap": true,
+	},
+	"include": ["./**/*"],
+	"references": [
+		{
+			"path": "../../tsconfig.json",
+		},
+	],
+}
