+++ packages/framework/tree-agent-types/src/test/typeFactory.spec.ts
@@ -5,12 +5,7 @@
 
 import { strict as assert } from "node:assert";
 
-import { SchemaFactory } from "@fluidframework/tree/internal";
-import { isTypeFactoryType } from "@fluidframework/tree-agent-types/alpha";
-
-import { typeFactory as tf } from "../treeAgentTypes.js";
-
-const sf = new SchemaFactory("test");
+import { typeFactory as tf, isTypeFactoryType } from "../treeAgentTypes.js";
 
 describe("type factories", () => {
 	describe("primitive types", () => {
@@ -146,22 +141,6 @@ describe("type factories", () => {
 			const innerType = readonlyType.innerType;
 			assert(isTypeFactoryType(innerType));
 			assert.equal(innerType._kind, "array");
-		});
-
-		it("creates instanceOf type with ObjectNodeSchema", () => {
-			class MyClass extends sf.object("MyClass", {}) {}
-			const instanceOfType = tf.instanceOf(MyClass);
-			assert(isTypeFactoryType(instanceOfType));
-			assert.equal(instanceOfType._kind, "instanceof");
-			assert.equal(instanceOfType.schema, MyClass);
-		});
-
-		it("throws error for instanceOf with non-ObjectNodeSchema", () => {
-			class ArraySchema extends sf.array("ArraySchema", sf.string) {}
-			assert.throws(
-				() => tf.instanceOf(ArraySchema),
-				/typeFactory\.instanceOf only supports ObjectNodeSchema-based schema classes/,
-			);
 		});
 
 		it("throws error for empty union", () => {
