+++ packages/framework/tree-agent/src/index.ts
@@ -40,53 +40,5 @@ export type {
 	Context,
 } from "./api.js";
 export { llmDefault } from "./utils.js";
-export {
-	buildFunc,
-	exposeMethodsSymbol,
-	type ArgsTuple,
-	type ExposedMethods,
-	type Arg,
-	type FunctionDef,
-	type MethodKeys,
-	type BindableSchema,
-	type Ctor,
-	type IExposedMethods,
-} from "./methodBinding.js";
-export type {
-	exposePropertiesSymbol,
-	PropertyDef,
-	ExposedProperties,
-	IExposedProperties,
-} from "./propertyBinding.js";
-
-export {
-	typeFactory,
-	isTypeFactoryType,
-} from "./treeAgentTypes.js";
-
-export type {
-	TypeFactoryType,
-	TypeFactoryTypeKind,
-	TypeFactoryString,
-	TypeFactoryNumber,
-	TypeFactoryBoolean,
-	TypeFactoryDate,
-	TypeFactoryVoid,
-	TypeFactoryUndefined,
-	TypeFactoryNull,
-	TypeFactoryUnknown,
-	TypeFactoryArray,
-	TypeFactoryPromise,
-	TypeFactoryObject,
-	TypeFactoryRecord,
-	TypeFactoryMap,
-	TypeFactoryTuple,
-	TypeFactoryUnion,
-	TypeFactoryIntersection,
-	TypeFactoryLiteral,
-	TypeFactoryOptional,
-	TypeFactoryReadonly,
-	TypeFactoryFunction,
-	TypeFactoryFunctionParameter,
-	TypeFactoryInstanceOf,
-} from "./treeAgentTypes.js";
+export { type BindableSchema } from "./methodBinding.js";
+export { typeFactory, type TypeFactoryInstanceOf } from "./treeAgentTypes.js";
