+++ packages/framework/tree-agent/src/propertyBinding.ts
@@ -5,88 +5,17 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type { TreeNodeSchema } from "@fluidframework/tree";
+import {
+	exposePropertiesSymbol,
+	PropertyDef,
+	isTypeFactoryType,
+	type Ctor,
+	type ExposedProperties,
+	type IExposedProperties,
+	type TypeFactoryType,
+} from "@fluidframework/tree-agent-types/internal";
 
-import type { BindableSchema, Ctor } from "./methodBinding.js";
-import type { TypeFactoryType } from "./treeAgentTypes.js";
-import { isTypeFactoryType } from "./treeAgentTypes.js";
-
-/**
- * A symbol used to expose properties to the LLM.
- * @alpha
- */
-export const exposePropertiesSymbol: unique symbol = Symbol.for(
-	"@fluidframework/tree-agent/exposeProperties",
-);
-
-/**
- * A property definition class that describes the structure of the property
- * @alpha
- */
-export class PropertyDef {
-	public constructor(
-		/**
-		 * The name of the property.
-		 */
-		public readonly name: string,
-		/**
-		 * Optional description of the property.
-		 */
-		public readonly description: string | undefined,
-		/**
-		 * The schema defining the property's type.
-		 */
-		public readonly schema: TypeFactoryType,
-		/**
-		 * Whether the property is readonly.
-		 */
-		public readonly readOnly: boolean,
-	) {}
-}
-
-/**
- * An interface for exposing properties of schema classes to an agent.
- * @alpha
- */
-export interface ExposedProperties {
-	/**
-	 * Expose a property with type factory type and metadata.
-	 */
-	exposeProperty<S extends BindableSchema & Ctor, K extends string>(
-		schema: S,
-		name: K,
-		def: { schema: TypeFactoryType; description?: string; readOnly?: boolean },
-	): void;
-
-	/**
-	 * Expose a property with type factory type (simple form).
-	 */
-	exposeProperty<S extends BindableSchema & Ctor, K extends string>(
-		schema: S,
-		name: K,
-		tfType: TypeFactoryType,
-	): void;
-}
-
-/**
- * An interface that SharedTree schema classes should implement to expose their properties to the LLM.
- *
- * @remarks
- * The `getExposedProperties` free function will cause the method here to be called on the class passed to it.
- *
- * @privateremarks
- * Implementing this interface correctly seems tricky?
- * To actually implement it in a way that satisfies TypeScript,
- * classes need to declare both a static version and an instance version of the method
- * (the instance one can just delegate to the static one).
- *
- * @alpha
- */
-export interface IExposedProperties {
-	/**
-	 * Static method that exposes properties of this schema class to an agent.
-	 */
-	[exposePropertiesSymbol]?(properties: ExposedProperties): void;
-}
+import type { BindableSchema } from "./methodBinding.js";
 
 class ExposedPropertiesI implements ExposedProperties {
 	private readonly properties: Record<string, PropertyDef> = {};
@@ -94,14 +23,14 @@ class ExposedPropertiesI implements ExposedProperties {
 
 	public constructor(private readonly schemaClass: BindableSchema) {}
 
-	public exposeProperty<S extends BindableSchema & Ctor, K extends string>(
+	public exposeProperty<S extends Ctor, K extends string>(
 		schema: S,
 		name: K,
 		defOrType:
 			| { schema: TypeFactoryType; description?: string; readOnly?: boolean }
 			| TypeFactoryType,
 	): void {
-		if (schema !== this.schemaClass) {
+		if ((schema as unknown) !== (this.schemaClass as unknown)) {
 			throw new Error('Must expose properties on the "this" schema class');
 		}
