+++ packages/framework/tree-agent/src/renderSchemaTypeScript.ts
@@ -13,14 +13,19 @@ import {
 } from "@fluidframework/tree/alpha";
 import { FieldKind, NodeKind, ValueSchema } from "@fluidframework/tree/internal";
 import type { SimpleLeafNodeSchema } from "@fluidframework/tree/internal";
+import {
+	isTypeFactoryType,
+	type FunctionWrapper,
+	type PropertyDef,
+	type TypeFactoryOptional,
+	type TypeFactoryType,
+} from "@fluidframework/tree-agent-types/internal";
 
-import type { BindableSchema, FunctionWrapper } from "./methodBinding.js";
+import type { BindableSchema } from "./methodBinding.js";
 import { getExposedMethods } from "./methodBinding.js";
 import { fluidHandleTypeName } from "./prompt.js";
-import { getExposedProperties, type PropertyDef } from "./propertyBinding.js";
+import { getExposedProperties } from "./propertyBinding.js";
 import { renderTypeFactoryTypeScript } from "./renderTypeFactoryTypeScript.js";
-import type { TypeFactoryOptional, TypeFactoryType } from "./treeAgentTypes.js";
-import { isTypeFactoryType } from "./treeAgentTypes.js";
 import {
 	IdentifierCollisionResolver,
 	getFriendlyName,
